+++ src/components/swap/AssetSelector.tsx
@@ -13,15 +13,16 @@ import {
   isSui,
   Asset,
   ChainAsset,
+  ChainData,
 } from "@gardenfi/orderbook";
-import { assetInfoStore, ChainData } from "../../store/assetInfoStore";
+import { assetInfoStore } from "../../store/assetInfoStore";
 import { BTC, swapStore } from "../../store/swapStore";
 import { IOType, network } from "../../constants/constants";
 import { modalStore } from "../../store/modalStore";
 import { ChainsTooltip } from "./ChainsTooltip";
 import { AvailableChainsSidebar } from "./AvailableChainsSidebar";
 import { AnimatePresence, motion } from "framer-motion";
-import { formatAmount } from "../../utils/utils";
+import { formatBalance } from "../../utils/utils";
 import { useEVMWallet } from "../../hooks/useEVMWallet";
 import { useBitcoinWallet } from "@gardenfi/wallet-connectors";
 import { useStarknetWallet } from "../../hooks/useStarknetWallet";
@@ -160,8 +161,7 @@ export const AssetSelector: FC<props> = ({ onClose }) => {
                     .toNumber()
                 : balance?.toNumber();
           const fiatBalance =
-            formattedBalance &&
-            (Number(formattedBalance) * Number(fiatRate)).toFixed(5);
+            formattedBalance && (formattedBalance * fiatRate).toFixed(5);
 
           return {
             asset,
@@ -340,7 +340,7 @@ export const AssetSelector: FC<props> = ({ onClose }) => {
                     <img
                       src={c.icon}
                       alt={c.name}
-                      className={`h-full max-h-5 w-full max-w-5 rounded-full`}
+                      className={`h-full max-h-5 w-full max-w-5`}
                     />
                     {hoveredChain === c.name && (
                       <ChainsTooltip
@@ -429,8 +429,8 @@ export const AssetSelector: FC<props> = ({ onClose }) => {
                               weight="regular"
                               className={`!text-mid-grey`}
                             >
-                              {formatAmount(
-                                Number(formattedBalance),
+                              {formatBalance(
+                                formattedBalance,
                                 0,
                                 Math.min(asset.decimals, BTC.decimals)
                               )}