+++ src/components/swap/AvailableChainsSidebar.tsx
@@ -1,4 +1,4 @@
-import { ChainData } from "../../store/assetInfoStore";
+import { ChainData } from "@gardenfi/orderbook";
 import {
   ArrowLeftIcon,
   GradientScroll,
@@ -118,7 +118,7 @@ export const AvailableChainsSidebar = ({
                           <img
                             src={c.icon}
                             alt={c.name}
-                            className={`h-5 w-5 rounded-full`}
+                            className={`h-5 w-5`}
                           />
                           <Typography
                             size={"h5"}